api/base/python/requirements.txt
fastapi[standard]
firebase-admin
python-dotenv
uvicorn